api/shrinker/requirements.txt
# Function dependencies, for example:
# package>=version
RandomWords==0.3.0
google-cloud-firestore==2.5.3